requirements.txt
streamlit
tensorflow
numpy
gdown
pillow